requirements.txt
PyJWT~=2.3.0
Flask~=3.0.0
flask-cors~=4.0.0

Werkzeug~=3.0.1
